split the function to the files
