+++ openstack/ces/v1/metricdata/show_event_data.go
@@ -1,0 +1,52 @@
+package metricdata
+
+import (
+	golangsdk "github.com/opentelekomcloud/gophertelekomcloud"
+	"github.com/opentelekomcloud/gophertelekomcloud/internal/extract"
+)
+
+type ShowEventDataOpts struct {
+	// Query the namespace of a service.
+	Namespace string `q:"namespace"`
+	// Specifies the dimension. For example, the ECS dimension is instance_id.
+	// For details about the dimensions corresponding to the monitoring metrics of each service,
+	// see the monitoring metrics description of the corresponding service in Services Interconnected with Cloud Eye.
+	//
+	// Specifies the dimension. A maximum of three dimensions are supported,
+	// and the dimensions are numbered from 0 in dim.{i}=key,value format.
+	// The key cannot exceed 32 characters and the value cannot exceed 256 characters.
+	Dim string `q:"dim.0"`
+	// Specifies the event type.
+	Type string `q:"type"`
+	// Specifies the start time of the query.
+	From string `q:"from"`
+	// Specifies the end time of the query.
+	To string `q:"to"`
+}
+
+func ListEventData(client *golangsdk.ServiceClient, opts ShowEventDataOpts) ([]EventDataInfo, error) {
+	q, err := golangsdk.BuildQueryString(&opts)
+	if err != nil {
+		return nil, err
+	}
+
+	// GET /V1.0/{project_id}/event-data
+	raw, err := client.Get(client.ServiceURL("event-data")+q.String(), nil, nil)
+	if err != nil {
+		return nil, err
+	}
+
+	var res []EventDataInfo
+	err = extract.IntoSlicePtr(raw.Body, &res, "datapoints")
+
+	return res, err
+}
+
+type EventDataInfo struct {
+	// Specifies the event type, for example, instance_host_info.
+	Type string `json:"type"`
+	// Specifies when the event is reported. It is a UNIX timestamp and the unit is ms.
+	Timestamp int64 `json:"timestamp"`
+	// Specifies the host configuration information.
+	Value string `json:"value"`
+}
